Add Statistical Significance (#52)
* binary tp, fp, tn, fn cols for statistical significance

* IP: statistical significance notebook testing binary tp, fp, tn, fn cols method in style of get_crosstabs for easy integration

* Calculate significance of differences between ref group and all other groups for label, score, tp, tn, fp, fn

* mask and move significance methods to bias class

* ID ref groups for significance methods

* Asignificance calculation, predefinied ref group bias method

* Add significance calculation to all bias methods

+++ src/tests/test_bias_report.py
@@ -56,7 +56,12 @@ def helper(input_filename, expected_filename, config_file):
             print('testing {} ...'.format(col))
 
             try:
-                if np.mean(combined_data[col + "_x"] - combined_data[col + "_y"]) > EPS:
+                # TypeError: numpy boolean subtract, the `-` operator, is
+                # deprecated, use the bitwise_xor, the `^` operator, or the
+                # logical_xor function instead.
+                # found online that casting as float64 will go around, but
+                # would like to get Jesse's take on best way to avoid issue.
+                if np.mean(combined_data[col + "_x"].astype("float64") - combined_data[col + "_y"].astype("float64")) > EPS:
                     exp_mean = np.mean(combined_data[col + "_x"])
                     aeq_mean = np.mean(combined_data[col + "_y"])
                     s += "{} fails: Expected {} on average, but aequitas returned {}\n".format(col, exp_mean,                                                                              aeq_mean)
